Rename go module to github.com/apache/spark/connector/connect/client/go/v_3_4
