+++ connector/connect/client/go/examples/spark-connect-example-spark-session/main.go
@@ -18,7 +18,7 @@ package main
 
 import (
 	"flag"
-	"github.com/apache/spark/go/v_3_4/spark/sql"
+	"github.com/apache/spark/connector/connect/client/go/v_3_4/spark/sql"
 	"log"
 )
 
@@ -62,7 +62,7 @@ func main() {
 	if err != nil {
 		log.Fatalf("Failed: %s", err.Error())
 	}
-	
+
 	for _, f := range schema.Fields {
 		log.Printf("Field in row: %s - %s", f.Name, f.DataType.TypeName())
 	}
