+++ connector/connect/client/go/spark/sql/dataframe.go
@@ -23,7 +23,7 @@ import (
 	"github.com/apache/arrow/go/v12/arrow"
 	"github.com/apache/arrow/go/v12/arrow/array"
 	"github.com/apache/arrow/go/v12/arrow/ipc"
-	proto "github.com/apache/spark/go/v_3_4/internal/generated"
+	proto "github.com/apache/spark/connector/connect/client/go/v_3_4/internal/generated"
 	"io"
 )
