+++ connector/connect/common/src/main/protobuf/spark/connect/base.proto
@@ -28,7 +28,7 @@ import "spark/connect/types.proto";
 
 option java_multiple_files = true;
 option java_package = "org.apache.spark.connect.proto";
-option go_package = "github.com/apache/spark/go/v_3_4/internal/generated";
+option go_package = "github.com/apache/spark/connector/connect/client/go/v_3_4/internal/generated";
 
 // A [[Plan]] is the structure that carries the runtime information for the execution from the
 // client to the server. A [[Plan]] can either be of the type [[Relation]] which is a reference
