+++ connector/connect/common/src/main/protobuf/spark/connect/catalog.proto
@@ -24,7 +24,7 @@ import "spark/connect/types.proto";
 
 option java_multiple_files = true;
 option java_package = "org.apache.spark.connect.proto";
-option go_package = "github.com/apache/spark/go/v_3_4/internal/generated";
+option go_package = "github.com/apache/spark/connector/connect/client/go/v_3_4/internal/generated";
 
 // Catalog messages are marked as unstable.
 message Catalog {
