+++ connector/connect/common/src/main/protobuf/spark/connect/commands.proto
@@ -26,7 +26,7 @@ package spark.connect;
 
 option java_multiple_files = true;
 option java_package = "org.apache.spark.connect.proto";
-option go_package = "github.com/apache/spark/go/v_3_4/internal/generated";
+option go_package = "github.com/apache/spark/connector/connect/client/go/v_3_4/internal/generated";
 
 // A [[Command]] is an operation that is executed by the server that does not directly consume or
 // produce a relational result.
