+++ connector/connect/common/src/main/protobuf/spark/connect/types.proto
@@ -21,7 +21,7 @@ package spark.connect;
 
 option java_multiple_files = true;
 option java_package = "org.apache.spark.connect.proto";
-option go_package = "github.com/apache/spark/go/v_3_4/internal/generated";
+option go_package = "github.com/apache/spark/connector/connect/client/go/v_3_4/internal/generated";
 
 // This message describes the logical [[DataType]] of something. It does not carry the value
 // itself but only describes it.
